+++ lib/services/pool/lib/util.ts
@@ -6,8 +6,6 @@ import {
     GqlPoolTokenBase,
     GqlPoolTokenLinear,
     GqlPoolTokenPhantomStable,
-    GqlPoolTokenPhantomStableNestedUnion,
-    GqlPoolTokenUnion,
     GqlPoolWeighted,
 } from '~/apollo/generated/graphql-codegen-generated';
 import { AmountHumanReadable, AmountScaledString, TokenAmountHumanReadable } from '~/lib/services/token/token-types';
@@ -17,7 +15,6 @@ import {
     oldBnumScale,
     oldBnumScaleAmount,
     oldBnumToBnum,
-    oldBnumToHumanReadable,
     oldBnumZero,
 } from '~/lib/services/pool/lib/old-big-number';
 import OldBigNumber from 'bignumber.js';
@@ -39,7 +36,6 @@ import { Contract } from '@ethersproject/contracts';
 import { networkConfig } from '~/lib/config/network-config';
 import VaultAbi from '~/lib/abi/VaultAbi.json';
 import { formatFixed } from '@ethersproject/bignumber';
-import { replaceWethWithEth } from '../../token/token-util';
 
 export function poolScaleAmp(amp: string): BigNumber {
     // amp is stored with 3 decimals of precision
@@ -572,78 +568,4 @@ export function poolGetNestedTokenEstimateForPoolTokenAmounts({
 
 export function tokenAmountsAllZero(tokenAmounts: TokenAmountHumanReadable[]) {
     return tokenAmounts.filter((amount) => parseFloat(amount.amount) === 0).length === tokenAmounts.length;
-}
-
-export function calculateBptOut(
-    bptTotalSupply: string,
-    amountIn: string,
-    balance: string,
-    decimals: number = 18,
-    priceRate: string = '1.0',
-) {
-    const balanceScaled = oldBnumScale(balance, decimals).times(priceRate);
-    const amountRatio = oldBnumScale(amountIn, decimals).div(balanceScaled);
-    const bptOut = oldBnumScale(bptTotalSupply, 18).times(amountRatio);
-
-    return bptOut;
-}
-
-export function calculateAmountIn(bptTotalSupply: string, bptIn: OldBigNumber, balance: string) {
-    const bptRatio = bptIn.div(oldBnumScale(bptTotalSupply, 18));
-    const amountIn = oldBnumScale(balance, 18).times(bptRatio);
-
-    // return human readable so we can convert a sum to the correct decimals again (alternative?)
-    return oldBnumToHumanReadable(amountIn);
-}
-
-export function getInvestTokenForLinearPoolToken(
-    userInvestTokenBalances: TokenAmountHumanReadable[],
-    poolToken: GqlPoolTokenUnion | GqlPoolTokenPhantomStableNestedUnion,
-) {
-    return userInvestTokenBalances.find((balance) => {
-        if (poolToken.__typename === 'GqlPoolTokenLinear') {
-            return !!poolToken.pool.tokens.find((nestedPoolToken) => nestedPoolToken.address === balance.address);
-        }
-    })!;
-}
-
-export function getBptOutForToken(
-    userInvestTokenBalances: TokenAmountHumanReadable[],
-    poolToken: GqlPoolTokenUnion | GqlPoolTokenPhantomStableNestedUnion,
-    bptTotalSupply: string,
-) {
-    const investToken =
-        poolToken.__typename === 'GqlPoolTokenLinear'
-            ? getInvestTokenForLinearPoolToken(userInvestTokenBalances, poolToken)
-            : userInvestTokenBalances.find(
-                  (balance) =>
-                      balance.address === poolToken.address ||
-                      balance.address === replaceWethWithEth(poolToken.address),
-              )!;
-
-    // for the old weighted boosted pools only the 'selected' token is in the userInvestBalances
-    // so we return null to filter out the other tokens in the invest service
-    if (!investToken) {
-        return null;
-    }
-
-    const decimals =
-        poolToken.__typename === 'GqlPoolTokenLinear'
-            ? poolToken.pool.tokens.find((token) => token.address === investToken.address)?.decimals
-            : poolToken.decimals;
-
-    return {
-        bptOut: calculateBptOut(
-            bptTotalSupply,
-            investToken.amount,
-            poolToken.balance,
-            poolToken.decimals,
-            poolToken.priceRate,
-        ),
-        token: {
-            ...investToken,
-            decimals,
-            balance: poolToken.balance,
-        },
-    };
 }
